+++ plugins/CopyEngine/Ultracopier-Spec/FileExistsDialog.cpp
@@ -96,7 +96,7 @@ FileExistsDialog::FileExistsDialog(QWidget *parent, INTERNALTYPEPATH source,
 #endif
         ui->label_source_modified->setVisible(true);
         ui->label_content_source_size->setVisible(true);
-        ui->label_content_source_size->setText(QString::fromStdString(facilityEngine->sizeToString(size)));
+        ui->label_content_source_size->setText(QString::fromStdString(facilityEngine->sizeToString((double)size)));
         ui->label_content_source_modified->setVisible(true);
         ui->label_content_source_modified->setText(QDateTime::fromMSecsSinceEpoch(mdate*1000).toString());
     }
@@ -145,7 +145,7 @@ FileExistsDialog::FileExistsDialog(QWidget *parent, INTERNALTYPEPATH source,
 #endif
         ui->label_destination_modified->setVisible(true);
         ui->label_content_destination_size->setVisible(true);
-        ui->label_content_destination_size->setText(QString::fromStdString(facilityEngine->sizeToString(size)));
+        ui->label_content_destination_size->setText(QString::fromStdString(facilityEngine->sizeToString((double)size)));
         ui->label_content_destination_modified->setVisible(true);
         ui->label_content_destination_modified->setText(QDateTime::fromMSecsSinceEpoch(mdate*1000).toString());
     }
